+++ packages/service-fetchq-task/src/on-fetchq-register-queue.js
@@ -1,0 +1,11 @@
+module.exports = (_, { getContext }) => {
+  const queueName = getContext('fetchq.task.queueName');
+  const queueSettings = getContext('fetchq.task.queueSettings');
+
+  return [
+    {
+      ...queueSettings,
+      name: queueName,
+    },
+  ];
+};