+++ packages/service-fetchq-task/src/on-fetchq-register-worker.js
@@ -1,0 +1,13 @@
+const taskWorker = require('./task-worker');
+
+module.exports = (_, { getContext }) => {
+  const queueName = getContext('fetchq.task.queueName');
+  const workerSettings = getContext('fetchq.task.workerSettings');
+  return [
+    {
+      ...workerSettings,
+      queue: queueName,
+      handler: taskWorker,
+    },
+  ];
+};